feat(dashboard): add derived calculation metadata and update utils
